% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_multiplets_seurat.R
\name{remove_multiplets_seurat}
\alias{remove_multiplets_seurat}
\title{Remove detected multiplets from a Seurat object}
\usage{
remove_multiplets_seurat(
  seurat_obj,
  multiplets,
  barcode_col = "barcode",
  classification_col = "our_classification"
)
}
\arguments{
\item{seurat_obj}{A Seurat object whose cell names (colnames) are barcodes.}

\item{multiplets}{Either (a) the data frame returned by
\code{detect_multiplets} (which has an \code{our_classification} column and
a barcode column), or (b) a character vector of multiplet barcodes to remove.}

\item{barcode_col}{Name of the barcode column when \code{multiplets} is a data
frame. Default "barcode".}

\item{classification_col}{Name of the classification column when
\code{multiplets} is a data frame. Default "our_classification".}
}
\value{
The Seurat object subset to the cells that are NOT multiplets.
}
\description{
Convenience function to subset a Seurat object to the singlet barcodes,
removing the barcodes that \code{\link{detect_multiplets}} flagged as
multiplets. This is a thin wrapper around Seurat's subsetting so users do not
have to match barcodes by hand.
}
\examples{
\dontrun{
  res <- detect_multiplets("gem_classification.csv", "out.csv")
  seu <- remove_multiplets_seurat(seu, res)
}
}
